Reimplement EditModById and implement TransformBinariesInModArchiveAsync

+++ DevilDaggersInfo.Web.BlazorWasm.Server/Services/Transient/ModFileSystemProcessor.cs
@@ -2,6 +2,7 @@ using DevilDaggersInfo.Core.Mod.Utils;
 using DevilDaggersInfo.Web.BlazorWasm.Server.Caches.ModArchives;
 using DevilDaggersInfo.Web.BlazorWasm.Server.InternalModels;
 using DevilDaggersInfo.Web.BlazorWasm.Shared.Utils;
+using System.Linq;
 
 namespace DevilDaggersInfo.Web.BlazorWasm.Server.Services.Transient;
 
@@ -9,11 +10,13 @@ public class ModFileSystemProcessor
 {
 	private readonly IFileSystemService _fileSystemService;
 	private readonly ModArchiveCache _modArchiveCache;
+	private readonly ModFileSystemAccessor _modFileSystemAccessor;
 
-	public ModFileSystemProcessor(IFileSystemService fileSystemService, ModArchiveCache modArchiveCache)
+	public ModFileSystemProcessor(IFileSystemService fileSystemService, ModArchiveCache modArchiveCache, ModFileSystemAccessor modFileSystemAccessor)
 	{
 		_fileSystemService = fileSystemService;
 		_modArchiveCache = modArchiveCache;
+		_modFileSystemAccessor = modFileSystemAccessor;
 	}
 
 	public async Task ProcessModBinaryUploadAsync(string modName, Dictionary<string, byte[]> binaries, List<FileSystemInformation> fileSystemInformation)
@@ -48,66 +51,41 @@ public class ModFileSystemProcessor
 		// We read and extract the .zip file we just created to validate it and to fill the cache if everything is OK.
 		byte[] zipBytes = IoFile.ReadAllBytes(zipFilePath);
 
-		try
-		{
-			List<ModBinaryCacheData> archive = _modArchiveCache.GetArchiveDataByBytes(modName, zipBytes).Binaries;
-			if (archive.Count == 0)
-				throw new InvalidModArchiveException("Mod archive does not contain any binaries.");
-
-			foreach (ModBinaryCacheData binary in archive)
-			{
-				if (binary.Chunks.Count == 0)
-					throw new InvalidModBinaryException($"Mod binary '{binary.Name}' does not contain any assets.");
-
-				if (!(binary.ModBinaryType is ModBinaryType.Audio or ModBinaryType.Dd))
-					throw new InvalidModBinaryException($"Mod binary '{binary.Name}' is a '{binary.ModBinaryType}' mod which is not allowed.");
-
-				string expectedPrefix = BinaryFileNameUtils.GetBinaryPrefix(binary.ModBinaryType, modName);
-
-				if (!binary.Name.StartsWith(expectedPrefix))
-					throw new InvalidModBinaryException($"Name of mod binary '{binary.Name}' must start with '{expectedPrefix}'.");
-
-				if (binary.Name.Length == expectedPrefix.Length)
-					throw new InvalidModBinaryException($"Name of mod binary '{binary.Name}' must not be equal to '{expectedPrefix}'.");
-			}
-		}
-		catch (InvalidModArchiveException)
-		{
-			if (IoFile.Exists(zipFilePath))
-				IoFile.Delete(zipFilePath);
-
-			throw;
-		}
-		catch (InvalidModBinaryException)
-		{
-			if (IoFile.Exists(zipFilePath))
-				IoFile.Delete(zipFilePath);
-
-			throw;
-		}
+		ValidateModArchiveOnDisk(modName, zipFilePath, zipBytes);
 
 		fileSystemInformation.Add(new($"File {_fileSystemService.FormatPath(zipFilePath)} ({FileSizeUtils.Format(zipBytes.Length)}) with binaries {string.Join(", ", binaries.Select(kvp => $"`{kvp.Key}`"))} was added.", FileSystemInformationType.Add));
 	}
 
-	public void ProcessModScreenshotUpload(string modName, Dictionary<string, byte[]> screenshots, List<FileSystemInformation> fileSystemInformation)
+	public async Task TransformBinariesInModArchiveAsync(string modName, List<string> binariesToDelete, Dictionary<string, byte[]> newBinaries, List<FileSystemInformation> fileSystemInformation)
 	{
-		string modScreenshotsDirectory = Path.Combine(_fileSystemService.GetPath(DataSubDirectory.ModScreenshots), modName);
-		Directory.CreateDirectory(modScreenshotsDirectory);
-		int i = 0;
-		foreach (KeyValuePair<string, byte[]> kvp in screenshots.OrderBy(kvp => kvp.Key))
+		if (binariesToDelete.Count == 0 && newBinaries.Count == 0)
+			return;
+
+		string zipFilePath = _modFileSystemAccessor.GetModArchivePath(modName);
+		using ZipArchive archive = ZipFile.Open(zipFilePath, ZipArchiveMode.Read);
+
+		Dictionary<string, byte[]> keptBinaries = new();
+		foreach (ZipArchiveEntry entry in archive.Entries)
 		{
-			if (!PngFileUtils.HasValidPngHeader(kvp.Value))
-			{
-				fileSystemInformation.Add(new($"File {kvp.Key} was skipped because it is not a valid PNG file.", FileSystemInformationType.Skip));
-				continue;
-			}
+			byte[] extractedContents = new byte[entry.Length];
 
-			string path = Path.Combine(modScreenshotsDirectory, $"{i:00}.png");
-			IoFile.WriteAllBytes(path, kvp.Value);
-			fileSystemInformation.Add(new($"File {_fileSystemService.FormatPath(path)} was added.", FileSystemInformationType.Add));
+			using Stream entryStream = entry.Open();
+			int readBytes = StreamUtils.ForceReadAllBytes(entryStream, extractedContents, 0, extractedContents.Length);
+			if (readBytes != extractedContents.Length)
+				throw new InvalidOperationException($"Reading all bytes from archived mod binary did not complete. {readBytes} out of {extractedContents.Length} bytes were read.");
 
-			i++;
+			keptBinaries.Add(entry.Name, extractedContents);
 		}
+
+		string? firstCollision = keptBinaries.Keys.FirstOrDefault(keptName => newBinaries.Any(kvp => kvp.Key == keptName));
+		if (firstCollision != null)
+			throw new InvalidModArchiveException($"Cannot append binary '{firstCollision}' to mod archive because it already contains a binary with the exact same name.");
+
+		if (IoFile.Exists(zipFilePath))
+			IoFile.Delete(zipFilePath);
+
+		Dictionary<string, byte[]> combinedBinaries = new List<Dictionary<string, byte[]>>() { keptBinaries, newBinaries }.SelectMany(dict => dict).ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
+		await ProcessModBinaryUploadAsync(modName, combinedBinaries, fileSystemInformation);
 	}
 
 	/// <summary>
@@ -115,6 +93,9 @@ public class ModFileSystemProcessor
 	/// </summary>
 	public void MoveModFilesAndClearCache(string newName, string currentName, List<FileSystemInformation> fileSystemInformation)
 	{
+		if (newName == currentName)
+			return;
+
 		string directory = _fileSystemService.GetPath(DataSubDirectory.Mods);
 		string oldPath = Path.Combine(directory, $"{currentName}.zip");
 		if (IoFile.Exists(oldPath))
@@ -192,6 +173,27 @@ public class ModFileSystemProcessor
 		}
 	}
 
+	public void ProcessModScreenshotUpload(string modName, Dictionary<string, byte[]> screenshots, List<FileSystemInformation> fileSystemInformation)
+	{
+		string modScreenshotsDirectory = Path.Combine(_fileSystemService.GetPath(DataSubDirectory.ModScreenshots), modName);
+		Directory.CreateDirectory(modScreenshotsDirectory);
+		int i = 0;
+		foreach (KeyValuePair<string, byte[]> kvp in screenshots.OrderBy(kvp => kvp.Key))
+		{
+			if (!PngFileUtils.HasValidPngHeader(kvp.Value))
+			{
+				fileSystemInformation.Add(new($"File {kvp.Key} was skipped because it is not a valid PNG file.", FileSystemInformationType.Skip));
+				continue;
+			}
+
+			string path = Path.Combine(modScreenshotsDirectory, $"{i:00}.png");
+			IoFile.WriteAllBytes(path, kvp.Value);
+			fileSystemInformation.Add(new($"File {_fileSystemService.FormatPath(path)} was added.", FileSystemInformationType.Add));
+
+			i++;
+		}
+	}
+
 	public void DeleteScreenshotsDirectory(string modName, List<FileSystemInformation> fileSystemInformation)
 	{
 		string screenshotsDirectory = Path.Combine(_fileSystemService.GetPath(DataSubDirectory.ModScreenshots), modName);
@@ -203,6 +205,47 @@ public class ModFileSystemProcessor
 		else
 		{
 			fileSystemInformation.Add(new($"Directory {_fileSystemService.FormatPath(screenshotsDirectory)} was not deleted because it does not exist.", FileSystemInformationType.NotFound));
+		}
+	}
+
+	private void ValidateModArchiveOnDisk(string modName, string zipFilePath, byte[] zipBytes)
+	{
+		try
+		{
+			List<ModBinaryCacheData> archive = _modArchiveCache.GetArchiveDataByBytes(modName, zipBytes).Binaries;
+			if (archive.Count == 0)
+				throw new InvalidModArchiveException("Mod archive does not contain any binaries.");
+
+			foreach (ModBinaryCacheData binary in archive)
+			{
+				if (binary.Chunks.Count == 0)
+					throw new InvalidModBinaryException($"Mod binary '{binary.Name}' does not contain any assets.");
+
+				if (!(binary.ModBinaryType is ModBinaryType.Audio or ModBinaryType.Dd))
+					throw new InvalidModBinaryException($"Mod binary '{binary.Name}' is a '{binary.ModBinaryType}' mod which is not allowed.");
+
+				string expectedPrefix = BinaryFileNameUtils.GetBinaryPrefix(binary.ModBinaryType, modName);
+
+				if (!binary.Name.StartsWith(expectedPrefix))
+					throw new InvalidModBinaryException($"Name of mod binary '{binary.Name}' must start with '{expectedPrefix}'.");
+
+				if (binary.Name.Length == expectedPrefix.Length)
+					throw new InvalidModBinaryException($"Name of mod binary '{binary.Name}' must not be equal to '{expectedPrefix}'.");
+			}
+		}
+		catch (InvalidModArchiveException)
+		{
+			if (IoFile.Exists(zipFilePath))
+				IoFile.Delete(zipFilePath);
+
+			throw;
+		}
+		catch (InvalidModBinaryException)
+		{
+			if (IoFile.Exists(zipFilePath))
+				IoFile.Delete(zipFilePath);
+
+			throw;
 		}
 	}
 }
